vector-store/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.8.0
pydantic-settings==2.1.0
weaviate-client==3.25.3
python-dotenv==1.0.0 